+++ src/components/organismus/SongSearchContainer.tsx
@@ -1,15 +1,10 @@
-import { FC, useEffect, useState, ChangeEvent } from 'react';
+import { FC, ChangeEvent } from 'react';
+import { useSongSearch } from 'hooks/useSongSearch';
 import { useNavigate } from 'react-router-dom';
-import { searchSongs } from 'service/api';
 import { SongSearchFormControl } from 'components/molecules/SongSearchFormControl';
-import Song from 'types/Song';
-
-const DEBOUNCING_DELAY = 500;
 
 const SongSearchConteiner: FC = () => {
-  const [keyword, setKeyword] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
-  const [songList, setSongList] = useState<Song[]>([]);
+  const [{ keyword, isLoading, songList }, { setKeyword }] = useSongSearch();
   const navigate = useNavigate();
 
   const handleChangeKeyword = (event: ChangeEvent<HTMLInputElement>) => {
@@ -26,30 +21,6 @@ const SongSearchConteiner: FC = () => {
     value: keyword,
     onChange: handleChangeKeyword,
   };
-
-  useEffect(() => {
-    if (!keyword) {
-      setSongList([]);
-
-      return;
-    }
-
-    const controller = new AbortController();
-    const { signal } = controller;
-    const search = async () => {
-      setIsLoading(true);
-      const response = await searchSongs({ word: keyword }, { signal });
-      setSongList(response.songs);
-      setIsLoading(false);
-    };
-
-    const timer = setTimeout(search, DEBOUNCING_DELAY);
-
-    return () => {
-      clearTimeout(timer);
-      controller.abort('search query is updated');
-    };
-  }, [keyword]);
 
   return (
     <>
